refactor: split Trade Market tab build and CtTransferList layout parts (#3878)
Extract trade_screen_market_tab_build.dart and CtTransferList mutation/layout
library parts so the largest remaining app shell widgets stay under file-size
targets; document the Market tab build part in repo-and-packages.md.

Refs #3878
