+++ app/lib/features/game/screens/trade/trade_screen.dart
@@ -94,6 +94,7 @@ part 'trade_screen_deal_book_rows.dart';
 part 'trade_screen_market_row.dart';
 part 'trade_screen_market_row_stepper.dart';
 part 'trade_screen_market_tab.dart';
+part 'trade_screen_market_tab_build.dart';
 part 'trade_screen_market_tab_cargo_header.dart';
 part 'trade_screen_market_tab_order_handlers.dart';
 part 'trade_screen_market_tab_catalog.dart';
