+++ app/lib/features/game/screens/trade/trade_screen_market_tab_build.dart
@@ -1,0 +1,252 @@
+// Market-tab `build` implementation for `_MarketTabContent`.
+// Split from `trade_screen_market_tab.dart` to keep each trade-screen part
+// under the repo file-size target (Refs #3878).
+
+part of 'trade_screen.dart';
+
+extension _MarketTabContentBuild on _MarketTabContent {
+  /// Visual dim factor applied to the Market tab body when the screen
+  /// is in observe mode (`canMutateViaUi == false`). Matches the
+  /// editorial-monocle conventions for read-only surfaces.
+  static const double observeModeOpacity = 0.7;
+
+  Widget buildMarketTabBody(BuildContext context, WidgetRef ref) {
+    final ThemeData theme = Theme.of(context);
+    final TextStyle nameStyle =
+        (theme.textTheme.titleSmall ?? const TextStyle(fontSize: 14))
+            .copyWith(color: EditorialMonoclePalette.accent);
+    final TextStyle priceStyle =
+        (theme.textTheme.titleSmall ?? const TextStyle(fontSize: 14))
+            .copyWith(color: EditorialMonoclePalette.accentBright);
+    final TextStyle volumeStyle =
+        (theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12))
+            .copyWith(color: EditorialMonoclePalette.muted);
+    final TextStyle quantityStyle =
+        (theme.textTheme.titleSmall ?? const TextStyle(fontSize: 14))
+            .copyWith(color: EditorialMonoclePalette.accentBright);
+    final TextStyle cargoIndicatorStyle =
+        (theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12))
+            .copyWith(color: EditorialMonoclePalette.accent);
+    final TextStyle cargoWarningStyle =
+        (theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12))
+            .copyWith(color: EditorialMonoclePalette.danger);
+
+    final _SectionedTradeableCommodities sectioned =
+        _tradeableCommoditiesByCategory();
+    final AppLocalizations l10n = appL10n(context);
+    final WorldMarketState market = game.worldMarketState;
+    final Orders orders = ref.watch(currentOrdersProvider);
+    final CurrentOrdersNotifier ordersNotifier =
+        ref.read(currentOrdersProvider.notifier);
+
+    int? readProjectedTreasuryDelta() {
+      try {
+        return ref.read(treasurySummaryProvider).projectedDelta;
+      } on Object {
+        return null;
+      }
+    }
+
+    final int tradeCargoCapacity = cargoHoldsForHomeFleet(game, playerId);
+    final int totalStagedBid = _totalStagedBidQuantity(orders, playerId);
+    final int remainingCargo = tradeCargoCapacity - totalStagedBid;
+    final int clampedRemaining = remainingCargo < 0 ? 0 : remainingCargo;
+    final bool warningVisible =
+        clampedRemaining == 0 && totalStagedBid > 0;
+
+    // Refs #3093 — sellable clamp slice. Compute the per-commodity offer
+    // cap and the player's already-staged offer quantities once per
+    // build so the rows below render `(headroom)` and clamp Offer
+    // toggles / `+` increments consistently. The helpers live in
+    // `colonizethis_logic` (pure functions) and are safe to call per
+    // frame. Industry-allocation reservations come from the player's
+    // current production assignments (derived from
+    // `productionDesiredOutputProvider`); subtracting them from the
+    // raw stockpile makes the sellable readout match
+    // `SPEC/game/world-market.md` § Per-commodity quantity cap. We
+    // `watch` the provider so allocation changes (e.g. the user
+    // returning from the Production screen) immediately refresh the
+    // sellable readouts without leaving the Market tab.
+    final Map<String, int> desiredOutputByRecipe =
+        ref.watch(productionDesiredOutputProvider);
+    final Map<CommodityId, int> productionInputConsumption =
+        _consumptionForDesiredOutput(desiredOutputByRecipe);
+    final Map<CommodityId, int> offerCap = offerCapByCommodityId(
+      game: game,
+      playerId: playerId,
+      productionInputConsumptionByCommodityId: productionInputConsumption,
+    );
+    final Map<CommodityId, int> stagedOffers =
+        stagedOfferQuantitiesByCommodityId(
+      orders: orders,
+      playerId: playerId,
+    );
+
+    // Refs #3546 — collapse the three verbatim per-section closure pairs into a
+    // single factory call. The shared per-build order context is bound once;
+    // the projected treasury delta is still read lazily per interaction inside
+    // the factory (behaviour-preserving).
+    final TradeSectionHandlers sectionHandlers = buildTradeSectionHandlers(
+      readProjectedTreasuryDelta: readProjectedTreasuryDelta,
+      handleDirectionChanged: ({
+        required CommodityId commodityId,
+        required TradeOrderType? next,
+        required int? projectedTreasuryDelta,
+      }) =>
+          _handleDirectionChanged(
+            ordersNotifier: ordersNotifier,
+            orders: orders,
+            productionInputConsumption: productionInputConsumption,
+            projectedTreasuryDelta: projectedTreasuryDelta,
+            commodityId: commodityId,
+            next: next,
+          ),
+      handleQuantityDelta: ({
+        required CommodityId commodityId,
+        required int delta,
+        required int? projectedTreasuryDelta,
+      }) =>
+          _handleQuantityDelta(
+            ordersNotifier: ordersNotifier,
+            orders: orders,
+            productionInputConsumption: productionInputConsumption,
+            projectedTreasuryDelta: projectedTreasuryDelta,
+            commodityId: commodityId,
+            delta: delta,
+          ),
+    );
+
+    // SingleChildScrollView + Column (instead of ListView.builder) so
+    // every commodity row is built up-front. Widget tests pin all 22
+    // tradeable rows by key without scrolling; the row count is bounded
+    // by the catalog size (22) so the eager build cost is negligible
+    // and the deterministic ordering survives Widgetbook stories that
+    // render the screen inside a non-scrollable container.
+    //
+    // Refs `#3093` — sectioned grouping slice. Rows are grouped by
+    // commodity category (Food → Raw Materials → Manufactured) under
+    // `CtSectionLabel` headers, mirroring the Production panel's
+    // Available subpanel so the two surfaces read consistently. Within
+    // each section, rows follow `CommodityCatalog.all` catalog order
+    // (no per-section alphabetical sort); this matches the Production
+    // panel's intra-section ordering.
+    final List<Widget> sectionWidgets = <Widget>[
+      ..._buildCommoditySectionWidgets(
+        sectionKey: TradeScreen.marketSectionFoodKey,
+        sectionLabel: l10n.production_food,
+        commodities: sectioned.food,
+        offerCap: offerCap,
+        stagedOffers: stagedOffers,
+        market: market,
+        orders: orders,
+        nameStyle: nameStyle,
+        priceStyle: priceStyle,
+        volumeStyle: volumeStyle,
+        quantityStyle: quantityStyle,
+        onDirectionChanged: sectionHandlers.onDirectionChanged,
+        onQuantityDelta: sectionHandlers.onQuantityDelta,
+      ),
+      ..._buildCommoditySectionWidgets(
+        sectionKey: TradeScreen.marketSectionRawMaterialsKey,
+        sectionLabel: l10n.production_rawMaterials,
+        commodities: sectioned.rawMaterials,
+        offerCap: offerCap,
+        stagedOffers: stagedOffers,
+        market: market,
+        orders: orders,
+        nameStyle: nameStyle,
+        priceStyle: priceStyle,
+        volumeStyle: volumeStyle,
+        quantityStyle: quantityStyle,
+        onDirectionChanged: sectionHandlers.onDirectionChanged,
+        onQuantityDelta: sectionHandlers.onQuantityDelta,
+        isFirstSection: false,
+      ),
+      ..._buildCommoditySectionWidgets(
+        sectionKey: TradeScreen.marketSectionManufacturedKey,
+        sectionLabel: l10n.production_manufactured,
+        commodities: sectioned.manufactured,
+        offerCap: offerCap,
+        stagedOffers: stagedOffers,
+        market: market,
+        orders: orders,
+        nameStyle: nameStyle,
+        priceStyle: priceStyle,
+        volumeStyle: volumeStyle,
+        quantityStyle: quantityStyle,
+        onDirectionChanged: sectionHandlers.onDirectionChanged,
+        onQuantityDelta: sectionHandlers.onQuantityDelta,
+        isFirstSection: false,
+      ),
+    ];
+
+    final Widget list = SingleChildScrollView(
+      key: TradeScreen.marketCommodityListKey,
+      padding: EdgeInsets.zero,
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        mainAxisSize: MainAxisSize.min,
+        children: sectionWidgets,
+      ),
+    );
+
+    final Widget header = _MarketTabCargoHeader(
+      clampedRemaining: clampedRemaining,
+      warningVisible: warningVisible,
+      cargoIndicatorStyle: cargoIndicatorStyle,
+      cargoWarningStyle: cargoWarningStyle,
+    );
+
+    final Widget body = Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        header,
+        // Flexible so the scrollable list still wins remaining height
+        // when the body is mounted inside a constrained column (the
+        // ancestor CtPanel + IndexedStack); when unconstrained it falls
+        // back to the natural intrinsic height.
+        Flexible(child: list),
+      ],
+    );
+
+    // Observe-mode (canMutateViaUi == false): wrap the **interactive**
+    // list in IgnorePointer + Opacity so the chips and stepper read as
+    // read-only, but leave the cargo indicator + warning header live
+    // (they're read-only telemetry that should still surface state).
+    if (!canEdit) {
+      return Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          header,
+          Flexible(
+            child: Opacity(
+              opacity: observeModeOpacity,
+              child: IgnorePointer(child: list),
+            ),
+          ),
+        ],
+      );
+    }
+    return body;
+  }
+
+  /// Pure helper: build the per-commodity production-input consumption
+  /// map for a desired-output snapshot. Extracted so `build` (which
+  /// `watch`es the provider) and the handlers (which `read` it) share
+  /// one normalisation path.
+  static Map<CommodityId, int> _consumptionForDesiredOutput(
+    Map<String, int> desiredOutputByRecipe,
+  ) {
+    if (desiredOutputByRecipe.isEmpty) {
+      return const <CommodityId, int>{};
+    }
+    final List<AssignedRecipe> assignments =
+        assignedRecipesFromDesiredOutput(desiredOutputByRecipe);
+    if (assignments.isEmpty) {
+      return const <CommodityId, int>{};
+    }
+    return productionInputConsumptionByCommodityIdForAssignments(assignments);
+  }
+}
